backend/requirements.txt
fastapi
uvicorn
gitpython
pytest
flake8
ruff
basedpyright
python-lsp-server
openai
python-dotenv
